Added persistence of administrative changes to switches and interfaces in case of abrupt failure
Today, the topology does not guarantee the persistence of administrative changes
in case of abrupt exit. This commit solves this problem by storing changes to the
administrative status of switches and interfaces in the NApp storehouse. He also
added a REST endpoint to restore these modifications in the event of an abrupt exit.