Update Link creation to handle exception (#127)
Today, an error occurs when the method _get_link_or_create is called with an empty endpoint. This commit changes this method, handling this exception and returning None when an exception is raised. In addition, this commit changes the method add_links to return if the link does not exist.

Co-authored-by: Humberto Diógenes <hdiogenes@gmail.com>