Load lldp and enabled/disabled status on Kytos bootstrap (#123)
* Restore of administrative statuses is now automatic

Before this commit, to restore administrative changes it was necessary
to use the REST Endpoint `restore`. This commit changes the
restore_network_status method, making this task automatic in the
kytos bootstrap.

* Add a timer to wait before restoring the administrative state of the interfaces

Wait for switch.interfaces to be filled. This avoids a false state of success
in restoring the entire administrative state of a switch's interfaces.

* Refactor the unit tests related to the method `restore_network_status`
* Add retry limits in waiting conditions
* Create a specific RestoreError exception

Co-authored-by: Antonio Francisco <antonio@ansp.br>
Co-authored-by: Gleyberson Andrade <gleybersonandrade@gmail.com>
Co-authored-by: Humberto Diógenes <hdiogenes@gmail.com>